requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.2
django-allauth==0.32.0
django-crispy-forms==1.6.1
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.5
lxml==4.1.0
oauthlib==2.0.2
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
urllib3==1.21.1
whitenoise==3.3.1
